train_requirements.txt
streamlit==1.37.0
pandas==2.1.4
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==2.2.2
